+++ packages/svelte/src/components/prosekit/index.ts
@@ -1,15 +1,14 @@
 import type { Editor } from '@prosekit/core'
-import type { SvelteComponent } from 'svelte'
+import type { Component, Snippet } from 'svelte'
 
 import Comp from './prosekit.svelte'
 
 export interface ProseKitProps {
   editor: Editor
+  children?: Snippet
 }
 
 /**
  * The root component for a ProseKit editor.
- *
- * @public
  */
-export const ProseKit = Comp as typeof SvelteComponent<any> as typeof SvelteComponent<ProseKitProps>
+export const ProseKit = Comp as Component<ProseKitProps>